refactor: rename generator crates with flavor suffixes and move to crates/generators/
- openapi-nexus-typescript → openapi-nexus-typescript-fetch
- openapi-nexus-go → openapi-nexus-go-http
- openapi-nexus-rust → openapi-nexus-rust-reqwest

Each generator now lives under crates/generators/ to accommodate future
language flavors (axios, hyper, reqwest-blocking, etc.). Workspace deps,
imports, justfile, docs, and the openapi-nexus-spec build.rs fixture path
are updated to match the new layout.
