+++ Cargo.toml
@@ -13,9 +13,9 @@ members = [
     "crates/openapi-nexus-plugin",
     "crates/openapi-nexus-spec",
     "crates/openapi-nexus-transforms",
-    "openapi-nexus-go",
-    "openapi-nexus-rust",
-    "openapi-nexus-typescript",
+    "crates/generators/openapi-nexus-go-http",
+    "crates/generators/openapi-nexus-rust-reqwest",
+    "crates/generators/openapi-nexus-typescript-fetch",
 ]
 
 [workspace.package]
@@ -71,14 +71,14 @@ openapi-nexus = { version = "0.0.5", path = "crates/openapi-nexus" }
 openapi-nexus-common = { version = "0.0.5", path = "crates/openapi-nexus-common" }
 openapi-nexus-config = { version = "0.0.5", path = "crates/openapi-nexus-config" }
 openapi-nexus-core = { version = "0.0.5", path = "crates/openapi-nexus-core" }
-openapi-nexus-go = { version = "0.0.5", path = "openapi-nexus-go" }
+openapi-nexus-go-http = { version = "0.0.5", path = "crates/generators/openapi-nexus-go-http" }
 openapi-nexus-ir = { version = "0.0.5", path = "crates/openapi-nexus-ir" }
 openapi-nexus-parser = { version = "0.0.5", path = "crates/openapi-nexus-parser" }
 openapi-nexus-plugin = { version = "0.0.5", path = "crates/openapi-nexus-plugin" }
-openapi-nexus-rust = { version = "0.0.5", path = "openapi-nexus-rust" }
+openapi-nexus-rust-reqwest = { version = "0.0.5", path = "crates/generators/openapi-nexus-rust-reqwest" }
 openapi-nexus-transforms = { version = "0.0.5", path = "crates/openapi-nexus-transforms" }
 openapi-nexus-spec = { version = "0.0.5", path = "crates/openapi-nexus-spec" }
-openapi-nexus-typescript = { version = "0.0.5", path = "openapi-nexus-typescript" }
+openapi-nexus-typescript-fetch = { version = "0.0.5", path = "crates/generators/openapi-nexus-typescript-fetch" }
 
 sigil-stitch = "0.2.1"
 
@@ -89,7 +89,7 @@ order = [
     "openapi-nexus-parser",
     "openapi-nexus-config",
     "openapi-nexus-core",
-    "openapi-nexus-typescript",
-    "openapi-nexus-go",
+    "openapi-nexus-typescript-fetch",
+    "openapi-nexus-go-http",
     "openapi-nexus",
 ]
