+++ crates/generators/openapi-nexus-go-http/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "openapi-nexus-typescript"
+name = "openapi-nexus-go-http"
 version.workspace = true
 edition.workspace = true
 description.workspace = true
@@ -12,8 +12,8 @@ publish = true
 [dependencies]
 heck.workspace = true
 serde.workspace = true
-sigil-stitch.workspace = true
 serde_json.workspace = true
+sigil-stitch.workspace = true
 snafu.workspace = true
 toml.workspace = true
 tracing.workspace = true
