+++ crates/generators/openapi-nexus-go-http/src/codegen.rs
