+++ crates/generators/openapi-nexus-go-http/src/config.rs
