+++ crates/generators/openapi-nexus-go-http/src/errors.rs
