+++ crates/generators/openapi-nexus-go-http/src/lib.rs
