+++ crates/generators/openapi-nexus-go-http/src/runtime.rs
