+++ crates/generators/openapi-nexus-go-http/src/runtime/auth.go.txt
