+++ crates/generators/openapi-nexus-go-http/src/runtime/client.go.txt
