+++ crates/generators/openapi-nexus-go-http/src/runtime/errors.go.txt
