+++ crates/generators/openapi-nexus-go-http/src/sigil_emit_api.rs
