+++ crates/generators/openapi-nexus-go-http/tests/golden_tests_go_http.rs
@@ -1,8 +1,8 @@
-//! Golden file tests for TypeScript Fetch generator
+//! Golden file tests for Go HTTP generator
 //!
-//! These tests compare generated TypeScript code against known-good golden files.
+//! These tests compare generated Go code against known-good golden files.
 //! To update golden files after intentional changes, run:
-//!   UPDATE_GOLDEN=1 cargo test --test golden_tests_typescript_fetch
+//!   UPDATE_GOLDEN=1 cargo test --test golden_tests_go_http
 
 use std::collections::HashMap;
 use std::env;
@@ -16,13 +16,14 @@ use tracing_test::traced_test;
 
 use openapi_nexus_core::traits::code_generator::CodeGenerator as _;
 use openapi_nexus_core::traits::file_writer::FileWriter;
+use openapi_nexus_go_http::GoHttpCodeGenerator;
 use openapi_nexus_spec::OpenApiV31Spec;
-use openapi_nexus_typescript::TypeScriptFetchCodeGenerator;
 
 /// Read a fixture file from various possible locations
 fn read_fixture(fixture_path: &str) -> String {
     let possible_paths = [
         Path::new("tests/fixtures").join(fixture_path),
+        Path::new("../../../tests/fixtures").join(fixture_path),
         Path::new("../../tests/fixtures").join(fixture_path),
         Path::new("../tests/fixtures").join(fixture_path),
     ];
@@ -37,7 +38,7 @@ fn read_fixture(fixture_path: &str) -> String {
 
 /// Get the golden directory path
 fn get_golden_dir() -> &'static Path {
-    Path::new("../tests/golden/typescript/typescript-fetch")
+    Path::new("../../../tests/golden/go/go-http")
 }
 
 /// Convert a generated filename to a golden filename by appending ".golden"
@@ -58,7 +59,7 @@ fn generate_files(
 ) -> Result<HashMap<String, String>, Box<dyn std::error::Error + Send + Sync>> {
     let openapi: OpenApiV31Spec = openapi_nexus_parser::parse_content_yaml_v31(spec_content)?;
 
-    let generator = TypeScriptFetchCodeGenerator::new(toml::value::Table::new());
+    let generator = GoHttpCodeGenerator::new(toml::value::Table::new());
     let generated_files = match generator.generate(&openapi) {
         Ok(files) => {
             println!("Successfully generated {} files", files.len());
@@ -137,10 +138,7 @@ fn test_golden_files(
     let generated = match generate_files(&spec_content) {
         Ok(files) => files,
         Err(e) => {
-            println!(
-                "Failed to generate TypeScript files for {}: {}",
-                spec_name, e
-            );
+            println!("Failed to generate Go files for {}: {}", spec_name, e);
             return Err(e);
         }
     };
@@ -286,7 +284,7 @@ fn show_diff(spec_name: &str, filename: &str, golden: &str, generated: &str) {
     );
 
     println!("To update golden files, run:");
-    println!("   UPDATE_GOLDEN=1 cargo test --test golden_tests_typescript_fetch");
+    println!("   UPDATE_GOLDEN=1 cargo test --test golden_tests_go_http");
 }
 
 /// Map of test case names to their fixture paths
@@ -300,12 +298,9 @@ fn get_golden_test_cases() -> HashMap<&'static str, &'static str> {
         ("duplicate-param-names", "valid/duplicate-param-names.yaml"),
         ("interface-with-enum-reference", "valid/interface-with-enum-reference.yaml"),
         ("minimal", "valid/minimal.yaml"),
-        ("multiple-similar-request-schemas", "valid/multiple-similar-request-schemas.yaml"),
         ("naming-conventions", "valid/naming-conventions.yaml"),
         ("server-object", "valid/server-object.yaml"),
 
-        ("recursive-json-all-optional-properties", "valid/recursive-json/all-optional-properties.yaml"),
-        ("recursive-json-all-optional-properties", "valid/recursive-json/all-optional-properties.yaml"),
         ("recursive-json-all-optional-properties", "valid/recursive-json/all-optional-properties.yaml"),
         ("recursive-json-array-of-inline-objects", "valid/recursive-json/array-of-inline-objects.yaml"),
         ("recursive-json-array-of-referenced-types", "valid/recursive-json/array-of-referenced-types.yaml"),
@@ -316,13 +311,11 @@ fn get_golden_test_cases() -> HashMap<&'static str, &'static str> {
         ("recursive-json-inline-object", "valid/recursive-json/inline-object.yaml"),
         ("recursive-json-inline-object-with-array", "valid/recursive-json/inline-object-with-array.yaml"),
         ("recursive-json-mixed-property-types", "valid/recursive-json/mixed-property-types.yaml"),
-        ("recursive-json-mixed-property-types", "valid/recursive-json/mixed-property-types.yaml"),
         ("recursive-json-nested-object-reference", "valid/recursive-json/nested-object-reference.yaml"),
         ("recursive-json-optional-array-of-inline-objects", "valid/recursive-json/optional-array-of-inline-objects.yaml"),
         ("recursive-json-optional-array-of-referenced-types", "valid/recursive-json/optional-array-of-referenced-types.yaml"),
         ("recursive-json-optional-inline-object", "valid/recursive-json/optional-inline-object.yaml"),
         ("recursive-json-optional-nested-object-reference", "valid/recursive-json/optional-nested-object-reference.yaml"),
-        ("recursive-json-primitive-array", "valid/recursive-json/primitive-array.yaml"),
         ("recursive-json-primitive-array", "valid/recursive-json/primitive-array.yaml"),
 
         ("type-aliases-complex-union", "valid/type-aliases/complex-union.yaml"),
@@ -383,7 +376,6 @@ generate_golden_tests! {
     test_duplicate_param_names_golden: "duplicate-param-names",
     test_interface_with_enum_reference_golden: "interface-with-enum-reference",
     test_minimal_golden: "minimal",
-    test_multiple_similar_request_schemas_golden: "multiple-similar-request-schemas",
     test_naming_conventions_golden: "naming-conventions",
     test_server_object_golden: "server-object",
 
@@ -398,12 +390,12 @@ generate_golden_tests! {
     test_recursive_json_inline_object_with_array_golden: "recursive-json-inline-object-with-array",
     test_recursive_json_mixed_property_types_golden: "recursive-json-mixed-property-types",
     test_recursive_json_nested_object_reference_golden: "recursive-json-nested-object-reference",
-
     test_recursive_json_optional_array_of_inline_objects_golden: "recursive-json-optional-array-of-inline-objects",
     test_recursive_json_optional_array_of_referenced_types_golden: "recursive-json-optional-array-of-referenced-types",
     test_recursive_json_optional_inline_object_golden: "recursive-json-optional-inline-object",
     test_recursive_json_optional_nested_object_reference_golden: "recursive-json-optional-nested-object-reference",
     test_recursive_json_primitive_array_golden: "recursive-json-primitive-array",
+
     test_type_aliases_complex_union_golden: "type-aliases-complex-union",
     test_type_aliases_intersection_allof_golden: "type-aliases-intersection-allof",
     test_type_aliases_intersection_with_nullable_reference_golden: "type-aliases-intersection-with-nullable-reference",
@@ -428,167 +420,4 @@ generate_golden_tests! {
     test_additional_properties_golden: "additional-properties",
 
     test_query_param_enum_golden: "query-param-enum",
-}
-
-// ===========================================================================
-// IR pipeline integration tests
-// ===========================================================================
-
-/// Test that the IR-based model generation produces output for a simple spec.
-#[test]
-fn test_ir_model_generation_produces_files() {
-    let yaml = r#"
-openapi: "3.1.0"
-info:
-  title: IR Test API
-  version: "1.0.0"
-components:
-  schemas:
-    User:
-      type: object
-      properties:
-        id:
-          type: integer
-        name:
-          type: string
-        email:
-          type: string
-      required:
-        - id
-        - name
-    Status:
-      type: string
-      enum:
-        - active
-        - inactive
-        - suspended
-"#;
-
-    let parsed = openapi_nexus_parser::parse_content_yaml(yaml).unwrap();
-    let ir = openapi_nexus_ir::lower::lower(parsed).unwrap();
-
-    let generator = TypeScriptFetchCodeGenerator::new(toml::value::Table::new());
-    let files = generator.generate_models_from_ir(&ir).unwrap();
-
-    // Should produce model files + index
-    assert!(
-        files.len() >= 3,
-        "Expected at least 3 files (User, Status, index), got {}",
-        files.len()
-    );
-
-    // Check we have model files with actual content
-    let filenames: Vec<&str> = files.iter().map(|f| f.filename.as_str()).collect();
-    assert!(
-        filenames.iter().any(|f| f.contains("User")),
-        "Expected a User model file, got: {:?}",
-        filenames
-    );
-    assert!(
-        filenames.iter().any(|f| f.contains("Status")),
-        "Expected a Status model file, got: {:?}",
-        filenames
-    );
-
-    // Verify User file contains interface definition
-    let user_file = files.iter().find(|f| f.filename.contains("User")).unwrap();
-    assert!(
-        user_file.content.contains("export interface User"),
-        "User file should contain 'export interface User', got:\n{}",
-        user_file.content
-    );
-    assert!(
-        user_file.content.contains("id:") || user_file.content.contains("id?:"),
-        "User file should contain 'id' property"
-    );
-
-    // Verify Status file contains enum definition
-    let status_file = files
-        .iter()
-        .find(|f| f.filename.contains("Status"))
-        .unwrap();
-    assert!(
-        status_file.content.contains("Status"),
-        "Status file should contain 'Status'"
-    );
-}
-
-/// Test that IR-based generation handles nullable types correctly.
-#[test]
-fn test_ir_model_nullable_types() {
-    let yaml = r#"
-openapi: "3.1.0"
-info:
-  title: Test
-  version: "1.0.0"
-components:
-  schemas:
-    Profile:
-      type: object
-      properties:
-        bio:
-          type:
-            - string
-            - "null"
-      required:
-        - bio
-"#;
-
-    let parsed = openapi_nexus_parser::parse_content_yaml(yaml).unwrap();
-    let ir = openapi_nexus_ir::lower::lower(parsed).unwrap();
-
-    let generator = TypeScriptFetchCodeGenerator::new(toml::value::Table::new());
-    let files = generator.generate_models_from_ir(&ir).unwrap();
-
-    let profile_file = files
-        .iter()
-        .find(|f| f.filename.contains("Profile"))
-        .unwrap();
-    // Nullable string in TS should contain "null" in the type
-    assert!(
-        profile_file.content.contains("null"),
-        "Profile file should contain null type for nullable property, got:\n{}",
-        profile_file.content
-    );
-}
-
-/// Test that IR-based generation handles refs (type aliases) correctly.
-#[test]
-fn test_ir_model_ref_alias() {
-    let yaml = r##"
-openapi: "3.1.0"
-info:
-  title: Test
-  version: "1.0.0"
-components:
-  schemas:
-    Pet:
-      type: object
-      properties:
-        name:
-          type: string
-    MyPet:
-      $ref: "#/components/schemas/Pet"
-"##;
-
-    let parsed = openapi_nexus_parser::parse_content_yaml(yaml).unwrap();
-    let ir = openapi_nexus_ir::lower::lower(parsed).unwrap();
-
-    let generator = TypeScriptFetchCodeGenerator::new(toml::value::Table::new());
-    let files = generator.generate_models_from_ir(&ir).unwrap();
-
-    // Should have at least Pet, MyPet, and index
-    assert!(
-        files.len() >= 3,
-        "Expected at least 3 files, got {}",
-        files.len()
-    );
-
-    let my_pet_file = files.iter().find(|f| f.filename.contains("MyPet")).unwrap();
-    // MyPet should be a type alias referencing Pet
-    assert!(
-        my_pet_file.content.contains("Pet"),
-        "MyPet file should reference Pet, got:\n{}",
-        my_pet_file.content
-    );
 }
