+++ crates/generators/openapi-nexus-rust-reqwest/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "openapi-nexus-rust"
+name = "openapi-nexus-rust-reqwest"
 version.workspace = true
 edition.workspace = true
 description.workspace = true
