+++ crates/generators/openapi-nexus-rust-reqwest/src/api_client.rs
