+++ crates/generators/openapi-nexus-rust-reqwest/src/emitter.rs
