+++ crates/generators/openapi-nexus-rust-reqwest/src/generator.rs
