+++ crates/generators/openapi-nexus-rust-reqwest/src/type_mapping.rs
