+++ crates/generators/openapi-nexus-typescript-fetch/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "openapi-nexus-go"
+name = "openapi-nexus-typescript-fetch"
 version.workspace = true
 edition.workspace = true
 description.workspace = true
@@ -12,8 +12,8 @@ publish = true
 [dependencies]
 heck.workspace = true
 serde.workspace = true
-serde_json.workspace = true
 sigil-stitch.workspace = true
+serde_json.workspace = true
 snafu.workspace = true
 toml.workspace = true
 tracing.workspace = true
