+++ crates/generators/openapi-nexus-typescript-fetch/src/codegen.rs
