+++ crates/generators/openapi-nexus-typescript-fetch/src/codegen/typescript_fetch_code_generator.rs
