+++ crates/generators/openapi-nexus-typescript-fetch/src/config.rs
