+++ crates/generators/openapi-nexus-typescript-fetch/src/config/module.rs
