+++ crates/generators/openapi-nexus-typescript-fetch/src/config/typescript_fetch_config.rs
