+++ crates/generators/openapi-nexus-typescript-fetch/src/errors.rs
