+++ crates/generators/openapi-nexus-typescript-fetch/src/lib.rs
