+++ crates/generators/openapi-nexus-typescript-fetch/src/project_files/README.md.txt
