+++ crates/generators/openapi-nexus-typescript-fetch/src/project_files/runtime.ts.txt
