+++ crates/generators/openapi-nexus-typescript-fetch/src/sigil_emit_api.rs
