+++ crates/generators/openapi-nexus-typescript-fetch/tests/sigil_api_scaffold.rs
@@ -11,13 +11,14 @@
 use std::fs;
 use std::path::Path;
 
-use openapi_nexus_typescript::sigil_emit_api::generate_api_files;
+use openapi_nexus_typescript_fetch::sigil_emit_api::generate_api_files;
 
 fn read_fixture(rel: &str) -> String {
     for base in [
         "tests/fixtures",
         "../tests/fixtures",
         "../../tests/fixtures",
+        "../../../tests/fixtures",
     ] {
         let p = Path::new(base).join(rel);
         if p.exists() {
