+++ crates/openapi-nexus/Cargo.toml
@@ -19,6 +19,6 @@ toml.workspace = true
 openapi-nexus-common.workspace = true
 openapi-nexus-config.workspace = true
 openapi-nexus-core.workspace = true
-openapi-nexus-go.workspace = true
+openapi-nexus-go-http.workspace = true
 openapi-nexus-parser.workspace = true
-openapi-nexus-typescript.workspace = true
+openapi-nexus-typescript-fetch.workspace = true
